feat: run npm i and build only in production via bash

+++ start.sh
@@ -1,16 +1,23 @@
 #!/usr/bin/env bash
 
-# Install new dependencies if any
-npm install
+# Check if the environment is production
+if [ "$NODE_ENV" = "production" ]; then
+    echo "Running in production mode"
 
-# Uninstall the current bcrypt modules
-npm uninstall bcrypt
+    # Install new dependencies if any
+    npm install
 
-# Install the bcrypt modules for the current environment
-npm install bcrypt
+    # Uninstall the current bcrypt modules
+    npm uninstall bcrypt
 
-# Build the TypeScript code
-npm run build
+    # Install the bcrypt modules for the current environment
+    npm install bcrypt
+
+    # Build the TypeScript code
+    npm run build
+else
+    echo "Not running in production mode, skipping npm install, build and bcrypt operations"
+fi
 
 echo "Starting API server"
